USB_PD now fully works

+++ ASS_V2/FUSB302.h
@@ -3,11 +3,13 @@
 #define _FUSB302_
 
 
+#include <Arduino.h>
+#include "FFat.h"
 #include <Wire.h>
 #include <cstdint>
 namespace Fusb302 {
 
-    uint8_t init(uint8_t sda, uint8_t scl, uint8_t i2cfreq);
+    uint8_t init(uint8_t sda, uint8_t scl, uint32_t i2cfreq);
 
 
 }
